lexai/api/tts.py
from fastapi import APIRouter, HTTPException
from pydantic import BaseModel
from typing import Optional
import logging

router = APIRouter()
logger = logging.getLogger(__name__)


class TTSRequest(BaseModel):
    text: str
    voice: Optional[str] = "default"
    language: Optional[str] = "en"
    speed: Optional[float] = 1.0


class TTSResponse(BaseModel):
    audio_url: str
    duration: float


@router.post("/synthesize", response_model=TTSResponse)
async def synthesize_speech(request: TTSRequest):
    try:
        # TODO: Implement Coqui TTS integration
        logger.info(f"TTS request received: {len(request.text)} characters")
        
        # Placeholder response
        return TTSResponse(
            audio_url="placeholder_audio_url",
            duration=0.0
        )
        
    except Exception as e:
        logger.error(f"Error in TTS synthesis: {e}")
        raise HTTPException(status_code=500, detail=str(e))