requirements.txt
# AI/ML Core
torch==2.7.0
torchaudio==2.7.0
transformers==4.52.3
accelerate==1.7.0
librosa==0.11.0
soundfile==0.13.1
pydub==0.25.1

# Text-to-Speech
TTS==0.22.0

# API Framework
fastapi==0.115.12
uvicorn==0.34.2
websockets==15.0.1
python-multipart==0.0.20

# Database
pymongo==4.13.0
motor==3.7.1

# Utilities
numpy==2.2.6
scipy==1.15.3
requests==2.32.3
python-dotenv==1.1.0
pydantic-settings==2.8.2

# Model optimization
GPUtil==1.4.0

# Audio processing
pyaudio==0.2.14

# Model hub
huggingface-hub==0.28.1
langdetect==1.0.9